requirements.txt
requests
pillow
django==3.2
python-dotenv==0.17.0